pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-csci005-nicolaslaub"
version = "0.1.3"
description = "A pirate-themed LLM chatbot powered by Groq"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "groq",
    "python-dotenv",
]

[project.scripts]
chat = "chat:repl"

[tool.setuptools]
py-modules = ["chat"]
packages = ["tools"]
